Clean-up in test/runners
