+++ tests/unit/bout_runners/executor/test_executor.py
@@ -1,0 +1,1 @@
+"""Contains unittests for the executor."""
